+++ packages/expo-ui/ios/Button/ButtonProps.swift
@@ -32,7 +32,8 @@ internal enum ButtonVariant: String, Enumerable {
   case plain
 }
 
-class ButtonProps: ExpoSwiftUI.ViewProps {
+class ButtonProps: ExpoSwiftUI.ViewProps, Observable {
+  required init() {}
   @Field var text: String = ""
   @Field var systemImage: String?
   @Field var buttonRole: ButtonRole? = .default