+++ packages/expo-ui/ios/SwitchView.swift
@@ -14,6 +14,7 @@ struct SwitchView: ExpoSwiftUI.View {
   @EnvironmentObject var props: SwitchProps
   @EnvironmentObject var shadowNodeProxy: ExpoSwiftUI.ShadowNodeProxy
   @State var checked: Bool = false
+  
 
   var body: some View {
     ExpoSwiftUI.AutoSizingStack(shadowNodeProxy: shadowNodeProxy, axis: .both) {
@@ -22,13 +23,17 @@ struct SwitchView: ExpoSwiftUI.View {
         if props.checked == newValue {
           return
         }
-        props.onCheckedChanged([
+        let payload = [
           "checked": newValue
-        ])
+        ]
+        props.onCheckedChanged(payload)
       })
       .onReceive(props.objectWillChange, perform: {
         checked = props.checked
       })
+      .onAppear() {
+        checked = props.checked
+      }
       #if !os(tvOS)
       .if(props.variant == "button") {
         $0.toggleStyle(.button)